Refactor core components into arianna_core package
