--- arianna/__init__.py
@@ -1,3 +1,0 @@
-from .http_client import HTTPClient
-
-__all__ = ["HTTPClient"]
