+++ tests/test_reasoning.py
@@ -13,10 +13,8 @@ from arianna_chain import (
     reason_loop,
     tree_reason_loop,
     multi_reason,
-    tokenizer,
-    validate_reasoning_tags,
-    ThoughtComplexityLogger,
 )
+from arianna_core import tokenizer, validate_reasoning_tags, ThoughtComplexityLogger
 
 
 def test_validate_reasoning_tags_valid() -> None:
